requirements.txt
asgiref==3.7.2
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2>=2.8.5,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
django-cors-headers
uwsgi>=2.0.19,<2.1